+++ src/scriptlets/scriptlets-list.js
@@ -61,5 +61,6 @@ export * from './remove-node-text';
 export * from './trusted-replace-node-text';
 export * from './evaldata-prune';
 export * from './trusted-prune-inbound-object';
+export * from './trusted-set-attr';
 export * from './spoof-css';
 export * from './call-nothrow';